:recycle:(wrappers): rename to methods and moved at the root
