+++ auto_tutorial_source/Segmentation/tutorial_muad_deep_en.py
@@ -33,7 +33,7 @@ from torchvision.utils import draw_segmentation_masks
 
 from torch_uncertainty import TUTrainer
 from torch_uncertainty.datamodules.segmentation import MUADDataModule
-from torch_uncertainty.models import deep_ensembles
+from torch_uncertainty.methods import deep_ensembles
 from torch_uncertainty.models.segmentation.unet import small_unet
 from torch_uncertainty.routines import SegmentationRoutine
 from torch_uncertainty.transforms import RepeatTarget
