+++ experiments/classification/mnist/configs/lenet/swag.yaml
@@ -24,7 +24,7 @@ trainer:
         check_finite: true
 routine:
   model:
-    class_path: torch_uncertainty.models.wrappers.SWAG
+    class_path: torch_uncertainty.methods.SWAG
     init_args:
       model:
         class_path: torch_uncertainty.models.classification.lenet
