+++ src/torch_uncertainty/methods/batch_ensemble.py
