+++ src/torch_uncertainty/methods/checkpoint_collector.py
