+++ src/torch_uncertainty/methods/stochastic.py
