+++ src/torch_uncertainty/methods/swa.py
