+++ src/torch_uncertainty/methods/swag.py
