+++ src/torch_uncertainty/models/__init__.py
@@ -24,17 +24,4 @@ from .classification import (
     wideresnet28x10,
 )
 from .depth import bts_resnet
-from .wrappers import (
-    EMA,
-    EPOCH_UPDATE_MODEL,
-    STEP_UPDATE_MODEL,
-    SWA,
-    SWAG,
-    BatchEnsemble,
-    CheckpointCollector,
-    StochasticModel,
-    Zero,
-    batch_ensemble,
-    deep_ensembles,
-    mc_dropout,
-)
+
