+++ src/torch_uncertainty/routines/pixel_regression.py
@@ -31,7 +31,7 @@ from torch_uncertainty.metrics import (
     SILog,
     ThresholdAccuracy,
 )
-from torch_uncertainty.models import (
+from torch_uncertainty.methods import (
     EPOCH_UPDATE_MODEL,
     STEP_UPDATE_MODEL,
 )
